requirements.txt
streamlit
opencv-python
tensorflow
keras
pillow
geocoder
folium
requests
streamlit-authenticator
